feat: 컴포넌트 명  변경 및  gql 쿼리 분리
