+++ front-end/mona_front/src/pages/Home/index.tsx
@@ -1,3 +1,3 @@
-import CardList from './components/CardList';
+import Home from './components/Home';
 
-export default CardList;
+export default Home;
